Add files via upload
Here is my lab7out files. We can use this as a base for the client, server, and database interfaces.

Here is the only dock I have from Dr Smith
Exisiting usernames were not  verified. -5
Your program allowed me to enter the same username twice - did not seem to check for an exisitng username. ﻿﻿﻿﻿﻿﻿﻿